+++ app/locales/en/translations.js
@@ -4,17 +4,13 @@ export default {
         abstract: 'Abstract',
         doi: 'DOI',
         tags: 'Tags',
-        open_science_framework: 'Open Science Framework',
+        openScienceFramework: 'Open Science Framework',
         title: 'Title',
         authors: 'Authors',
         license: 'License',
         none: 'None',
         settings: 'Settings',
         moderation: 'Moderation',
-        workflows: {
-            pre_moderation: 'pre-moderation',
-            post_moderation: 'post-moderation',
-        },
     },
     documentType: {
         default: {
@@ -53,21 +49,21 @@ export default {
             title: 'Moderate your collection',
             description: `At last, the ability to manage what scholarly works are displayed with your branding.
             No more working on off-topic papers confusing loyal researchers.`,
-            list_1: 'See all submissions in one place.',
-            list_2: 'Provide feedback to authors.',
-            list_3: 'Manage collections settings.',
+            list1: 'See all submissions in one place.',
+            list2: 'Provide feedback to authors.',
+            list3: 'Manage collections settings.',
         },
         workflow: {
             title: 'Choose your workflow',
-            list_1: 'Have greater control over what material is publicly available by choosing pre-moderation.',
-            list_2: 'Keep the time delay between submission and public access minimal by choosing post-moderation.',
+            list1: 'Have greater control over what material is publicly available by choosing pre-moderation.',
+            list2: 'Keep the time delay between submission and public access minimal by choosing post-moderation.',
             figure: {
-                pre_moderation: 'Pre-moderation',
-                post_moderation: 'Post-moderation',
+                preModeration: 'Pre-moderation',
+                postModeration: 'Post-moderation',
             },
         },
     },
-    provider_settings: {
+    providerSettings: {
         reviewsWorkflow: {
             title: 'Moderation Type',
             description: '',
@@ -112,14 +108,14 @@ export default {
         },
     },
     settings: {
-        not_editable: 'Moderation settings can only be changed by an OSF administrator. Contact support+{{provider.id}}@osf.io for assistance.',
+        notEditable: 'Moderation settings can only be changed by an OSF administrator. Contact support+{{provider.id}}@osf.io for assistance.',
     },
     setup: {
         start: 'Start Moderating',
         which: 'Which provider would you like to set up first?',
-        multiple_providers: 'You\'re an Admin for Multiple Providers',
-        choose_settings: 'Choose moderation settings for {{provider.name}}',
-        once_finalized: 'Once finalized, moderation settings can only be changed by an OSF administrator.',
+        multipleProviders: 'You\'re an Admin for Multiple Providers',
+        chooseSettings: 'Choose moderation settings for {{provider.name}}',
+        onceFinalized: 'Once finalized, moderation settings can only be changed by an OSF administrator.',
         finalize: 'Finalize Settings',
         error: {
             message: 'Unable to complete the setup of {{provider.name}}. Please contact support@osf.io.',
@@ -146,89 +142,88 @@ export default {
     },
     content: {
         header: {
-            last_edited: 'Last edited',
+            lastEdited: 'Last edited',
         },
-        date_label: {
-            created_on: 'Created on',
-            submitted_on: 'Submitted on',
+        dateLabel: {
+            createdOn: 'Created on',
+            submittedOn: 'Submitted on',
         },
         share: {
             download: 'Download',
             downloads: 'Downloads',
-            download_file: 'Download file',
-            download_preprint: 'Download {{provider.type.singular}}',
+            downloadFile: 'Download file',
+            downloadPreprint: 'Download {{provider.type.singular}}',
         },
-        see_more: 'See more',
-        see_less: 'See less',
+        seeMore: 'See more',
+        seeLess: 'See less',
         version: 'Version',
-        preprint_doi: '{{provider.type.singularCapitalized}} DOI',
-        preprint_pending_doi: 'DOI created after moderator approval',
-        article_doi: 'Peer-reviewed Publication DOI',
+        preprintDOI: '{{provider.type.singularCapitalized}} DOI',
+        preprintPendingDOI: 'DOI created after moderator approval',
+        articleDOI: 'Peer-reviewed Publication DOI',
         citations: 'Citations',
         disciplines: 'Disciplines',
-        project_button: {
+        projectButton: {
             paragraph: 'The project for this paper is available on the OSF.',
             button: 'Visit project',
-            edit_preprint: 'Edit {{provider.type.singular}}',
         },
     },
     components: {
-        'action-feed': {
-            no_actions: 'No recent activity.',
-            see_more: 'See more',
-            error_loading: 'Error fetching more events.',
+        actionFeed: {
+            noActions: 'No recent activity.',
+            seeMore: 'See more',
+            errorLoading: 'Error fetching more events.',
         },
-        'action-feed-entry': {
-            action_message: {
+        actionFeedEntry: {
+            actionMessage: {
                 submit: 'submitted a {{documentType}} to {{providerName}}',
                 accept: 'accepted a {{documentType}} in {{providerName}}',
                 reject: 'rejected a {{documentType}} from {{providerName}}',
                 edit_comment: 'edited the comment for a {{documentType}} in {{providerName}}',
             },
         },
-        'dashboard-sidebar': {
+        dashboardSidebar: {
             providers: 'Providers',
-            set_up: 'Set up moderation',
+            setUp: 'Set up moderation',
         },
-        'error-page': {
+        errorPage: {
             title: {
-                not_found: 'Page not found',
-                not_authenticated: 'Not logged in',
+                notFound: 'Page not found',
+                notAauthenticated: 'Not logged in',
                 forbidden: 'Forbidden',
-                not_setup: 'Moderation not enabled',
+                notSetup: 'Moderation not enabled',
             },
             details: {
-                not_found: 'The page you were looking for is not found on the OSF Preprints service.',
-                not_authenticated: 'You must be logged in to view this page.',
+                notFound: 'The page you were looking for is not found on the OSF Preprints service.',
+                notAuthenticated: 'You must be logged in to view this page.',
                 forbidden: 'You do not have the permissions to view this page.',
-                not_setup: 'An admin needs to set up moderation for this preprint provider.',
+                notSetup: 'An admin needs to set up moderation for this preprint provider.',
             },
             report: 'If this should not have occurred and the issue persists, please report it to',
-            go_to: 'Go to OSF Preprints',
+            goTo: 'Go to OSF Preprints',
         },
-        'moderation-list': {
+        moderationList: {
             newest: 'Newest',
             oldest: 'Oldest',
             pending: 'Pending',
             accepted: 'Accepted',
             rejected: 'Rejected',
             sort: 'Sort',
-            no_submissions: 'No submissions.',
+            noSubmissions: 'No submissions.',
         },
-        'moderation-list-row': {
+        moderationListRow: {
             submission: {
-                submitted_on: 'submitted on',
-                was_accepted_on: 'was accepted on',
-                was_rejected_on: 'was rejected on',
-                submitted: 'submitted',
-                was_accepted: 'was accepted',
-                was_rejected: 'was rejected',
-                by: 'by',
-                submission_by: 'submission by',
+                submittedOn: 'submitted on {{timeDate}} by',
+                submitted: 'submitted {{timeDate}} by',
+                acceptedOn: 'accepted on {{timeDate}} by {{moderatorName}}',
+                accepted: 'accepted {{timeDate}} by {{moderatorName}}',
+                acceptedAutomaticallyOn: 'accepted automatically on {{timeDate}}',
+                acceptedAutomatically: 'accepted automatically {{timeDate}}',
+                rejectedOn: 'rejected on {{timeDate}} by {{moderatorName}}',
+                rejected: 'rejected {{timeDate}} by {{moderatorName}}',
             },
         },
-        'preprint-status-banner': {
-            recent_activity: {
+        preprintStatusBanner: {
+            recentActivity: {
                 pending: 'submitted this {{provider.type.singular}} on',
                 accepted: 'accepted this {{provider.type.singular}} on',
                 rejected: 'rejected this {{provider.type.singular}} on',
@@ -238,8 +233,8 @@ export default {
                 },
             },
             message: {
-                pending_pre: 'not publicly available or searchable until approved by a moderator',
-                pending_post: 'publicly available and searchable but is subject to removal by a moderator',
+                pendingPre: 'not publicly available or searchable until approved by a moderator',
+                pendingPost: 'publicly available and searchable but is subject to removal by a moderator',
                 accepted: 'publicly available and searchable',
                 rejected: 'not publicly available or searchable',
             },
@@ -248,21 +243,21 @@ export default {
             rejected: 'rejected',
             loading: 'Loading...',
             decision: {
-                make_decision: 'Make decision',
-                modify_decision: 'Modify decision',
+                makeDecision: 'Make decision',
+                modifyDecision: 'Modify decision',
                 header: {
-                    submit_decision: 'Submit your decision',
-                    modify_decision: 'Modify your decision',
+                    submitDecision: 'Submit your decision',
+                    modifyDecision: 'Modify your decision',
                 },
                 moderator: 'Moderator',
                 base: 'This {{provider.type.singular}} is',
                 btn: {
-                    submit_decision: 'Submit decision',
-                    modify_decision: 'Modify decision',
+                    submitDecision: 'Submit decision',
+                    modifyDecision: 'Modify decision',
                     update_comment: 'Update comment',
                 },
-                comment_placeholder: 'Explain the reasoning behind your decision (optional)',
-                comment_length_error: 'Comment is {{difference}} character(s) too long (maximum is {{limit}}).',
+                commentPlaceholder: 'Explain the reasoning behind your decision (optional)',
+                commentLengthError: 'Comment is {{difference}} character(s) too long (maximum is {{limit}}).',
                 accept: {
                     label: 'Accept',
                     pre: 'Submission will appear in search results and be made public.',